update unit test for adc controller to test volume migration
