+++ pkg/controller/volume/attachdetach/testing/testvolumespec.go
@@ -18,8 +18,6 @@ package testing
 
 import (
 	"fmt"
-	"sync"
-	"time"
 
 	v1 "k8s.io/api/core/v1"
 	storagev1 "k8s.io/api/storage/v1"
@@ -32,8 +30,6 @@ import (
 	"k8s.io/kubernetes/pkg/volume"
 	"k8s.io/kubernetes/pkg/volume/util"
 )
-
-const TestPluginName = "kubernetes.io/testPlugin"
 
 // GetTestVolumeSpec returns a test volume spec
 func GetTestVolumeSpec(volumeName string, diskName v1.UniqueVolumeName) *volume.Spec {
@@ -93,7 +89,7 @@ func CreateTestClient() *fake.Clientset {
 							VolumeMounts: []v1.VolumeMount{
 								{
 									Name:      "volumeMountName",
-									ReadOnly:  false,
+									ReadOnly:  true,
 									MountPath: "/mnt",
 								},
 							},
@@ -104,9 +100,10 @@ func CreateTestClient() *fake.Clientset {
 							Name: "volumeName",
 							VolumeSource: v1.VolumeSource{
 								GCEPersistentDisk: &v1.GCEPersistentDiskVolumeSource{
-									PDName:   "pdName",
-									FSType:   "ext4",
-									ReadOnly: false,
+									PDName: "pdName",
+									FSType: "ext4",
+									// Make the translated volume allow Multi-Attach.
+									ReadOnly: true,
 								},
 							},
 						},
@@ -124,26 +121,6 @@ func CreateTestClient() *fake.Clientset {
 		pod := createAction.GetObject().(*v1.Pod)
 		extraPods.Items = append(extraPods.Items, *pod)
 		return true, createAction.GetObject(), nil
-	})
-	fakeClient.AddReactor("list", "csinodes", func(action core.Action) (handled bool, ret runtime.Object, err error) {
-		obj := &storagev1.CSINodeList{}
-		nodeNamePrefix := "mynode"
-		for i := 0; i < 5; i++ {
-			var nodeName string
-			if i != 0 {
-				nodeName = fmt.Sprintf("%s-%d", nodeNamePrefix, i)
-			} else {
-				// We want also the "mynode" node since all the testing pods live there
-				nodeName = nodeNamePrefix
-			}
-			csiNode := storagev1.CSINode{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: nodeName,
-				},
-			}
-			obj.Items = append(obj.Items, csiNode)
-		}
-		return true, obj, nil
 	})
 	nodes = &v1.NodeList{}
 	nodeNamePrefix := "mynode"
@@ -171,6 +148,23 @@ func CreateTestClient() *fake.Clientset {
 	fakeClient.AddReactor("list", "nodes", func(action core.Action) (handled bool, ret runtime.Object, err error) {
 		obj := &v1.NodeList{}
 		obj.Items = append(obj.Items, nodes.Items...)
+		return true, obj, nil
+	})
+	fakeClient.AddReactor("list", "csinodes", func(action core.Action) (handled bool, ret runtime.Object, err error) {
+		obj := &storagev1.CSINodeList{}
+		for _, node := range nodes.Items {
+			csiNode := storagev1.CSINode{
+				ObjectMeta: metav1.ObjectMeta{
+					Annotations: map[string]string{
+						// All the in-tree plugins have been migrated to CSI since v1.27.
+						// So hardcoding the migrated plugins here.
+						"storage.alpha.kubernetes.io/migrated-plugins": "kubernetes.io/aws-ebs,kubernetes.io/azure-disk,kubernetes.io/azure-file,kubernetes.io/cinder,kubernetes.io/gce-pd,kubernetes.io/vsphere-volume",
+					},
+					Name: node.Name,
+				},
+			}
+			obj.Items = append(obj.Items, csiNode)
+		}
 		return true, obj, nil
 	})
 	volumeAttachments = &storagev1.VolumeAttachmentList{}
@@ -235,7 +229,7 @@ func NewPodWithVolume(podName, volumeName, nodeName string) *v1.Pod {
 					VolumeMounts: []v1.VolumeMount{
 						{
 							Name:      "volumeMountName",
-							ReadOnly:  false,
+							ReadOnly:  true,
 							MountPath: "/mnt",
 						},
 					},
@@ -246,9 +240,10 @@ func NewPodWithVolume(podName, volumeName, nodeName string) *v1.Pod {
 					Name: volumeName,
 					VolumeSource: v1.VolumeSource{
 						GCEPersistentDisk: &v1.GCEPersistentDiskVolumeSource{
-							PDName:   "pdName",
-							FSType:   "ext4",
-							ReadOnly: false,
+							PDName: "pdName",
+							FSType: "ext4",
+							// Make the translated volume allow Multi-Attach.
+							ReadOnly: true,
 						},
 					},
 				},
@@ -347,242 +342,4 @@ func attachVolumeToNode(nodes *v1.NodeList, volumeName, nodeName string, inUse b
 	if inUse {
 		node.Status.VolumesInUse = append(node.Status.VolumesInUse, uniqueVolumeName)
 	}
-}
-
-type TestPlugin struct {
-	ErrorEncountered  bool
-	attachedVolumeMap map[string][]string
-	detachedVolumeMap map[string][]string
-	pluginLock        *sync.RWMutex
-}
-
-func (plugin *TestPlugin) Init(host volume.VolumeHost) error {
-	return nil
-}
-
-func (plugin *TestPlugin) GetPluginName() string {
-	return TestPluginName
-}
-
-func (plugin *TestPlugin) GetVolumeName(spec *volume.Spec) (string, error) {
-	plugin.pluginLock.Lock()
-	defer plugin.pluginLock.Unlock()
-	if spec == nil {
-		plugin.ErrorEncountered = true
-		return "", fmt.Errorf("GetVolumeName called with nil volume spec")
-	}
-	if spec.Volume != nil {
-		return spec.Name(), nil
-	} else if spec.PersistentVolume != nil {
-		if spec.PersistentVolume.Spec.PersistentVolumeSource.GCEPersistentDisk != nil {
-			return spec.PersistentVolume.Spec.PersistentVolumeSource.GCEPersistentDisk.PDName, nil
-		} else if spec.PersistentVolume.Spec.PersistentVolumeSource.NFS != nil {
-			return spec.PersistentVolume.Spec.PersistentVolumeSource.NFS.Server, nil
-		} else if spec.PersistentVolume.Spec.PersistentVolumeSource.RBD != nil {
-			return spec.PersistentVolume.Spec.PersistentVolumeSource.RBD.RBDImage, nil
-		}
-		return "", fmt.Errorf("GetVolumeName called with unexpected PersistentVolume: %v", spec)
-	} else {
-		return "", nil
-	}
-}
-
-func (plugin *TestPlugin) CanSupport(spec *volume.Spec) bool {
-	plugin.pluginLock.Lock()
-	defer plugin.pluginLock.Unlock()
-	if spec == nil {
-		plugin.ErrorEncountered = true
-	}
-	return true
-}
-
-func (plugin *TestPlugin) RequiresRemount(spec *volume.Spec) bool {
-	return false
-}
-
-func (plugin *TestPlugin) NewMounter(spec *volume.Spec, podRef *v1.Pod, opts volume.VolumeOptions) (volume.Mounter, error) {
-	plugin.pluginLock.Lock()
-	defer plugin.pluginLock.Unlock()
-	if spec == nil {
-		plugin.ErrorEncountered = true
-	}
-	return nil, nil
-}
-
-func (plugin *TestPlugin) NewUnmounter(name string, podUID types.UID) (volume.Unmounter, error) {
-	return nil, nil
-}
-
-func (plugin *TestPlugin) ConstructVolumeSpec(volumeName, mountPath string) (volume.ReconstructedVolume, error) {
-	fakeVolume := &v1.Volume{
-		Name: volumeName,
-		VolumeSource: v1.VolumeSource{
-			GCEPersistentDisk: &v1.GCEPersistentDiskVolumeSource{
-				PDName:   "pdName",
-				FSType:   "ext4",
-				ReadOnly: false,
-			},
-		},
-	}
-	return volume.ReconstructedVolume{
-		Spec: volume.NewSpecFromVolume(fakeVolume),
-	}, nil
-}
-
-func (plugin *TestPlugin) NewAttacher() (volume.Attacher, error) {
-	attacher := testPluginAttacher{
-		ErrorEncountered:  &plugin.ErrorEncountered,
-		attachedVolumeMap: plugin.attachedVolumeMap,
-		pluginLock:        plugin.pluginLock,
-	}
-	return &attacher, nil
-}
-
-func (plugin *TestPlugin) NewDeviceMounter() (volume.DeviceMounter, error) {
-	return plugin.NewAttacher()
-}
-
-func (plugin *TestPlugin) NewDetacher() (volume.Detacher, error) {
-	detacher := testPluginDetacher{
-		detachedVolumeMap: plugin.detachedVolumeMap,
-		pluginLock:        plugin.pluginLock,
-	}
-	return &detacher, nil
-}
-
-func (plugin *TestPlugin) CanAttach(spec *volume.Spec) (bool, error) {
-	return true, nil
-}
-
-func (plugin *TestPlugin) CanDeviceMount(spec *volume.Spec) (bool, error) {
-	return true, nil
-}
-
-func (plugin *TestPlugin) NewDeviceUnmounter() (volume.DeviceUnmounter, error) {
-	return plugin.NewDetacher()
-}
-
-func (plugin *TestPlugin) GetDeviceMountRefs(deviceMountPath string) ([]string, error) {
-	return []string{}, nil
-}
-
-func (plugin *TestPlugin) SupportsMountOption() bool {
-	return false
-}
-
-func (plugin *TestPlugin) SupportsBulkVolumeVerification() bool {
-	return false
-}
-
-func (plugin *TestPlugin) SupportsSELinuxContextMount(spec *volume.Spec) (bool, error) {
-	return false, nil
-}
-
-func (plugin *TestPlugin) GetErrorEncountered() bool {
-	plugin.pluginLock.RLock()
-	defer plugin.pluginLock.RUnlock()
-	return plugin.ErrorEncountered
-}
-
-func (plugin *TestPlugin) GetAttachedVolumes() map[string][]string {
-	plugin.pluginLock.RLock()
-	defer plugin.pluginLock.RUnlock()
-	ret := make(map[string][]string)
-	for nodeName, volumeList := range plugin.attachedVolumeMap {
-		ret[nodeName] = make([]string, len(volumeList))
-		copy(ret[nodeName], volumeList)
-	}
-	return ret
-}
-
-func (plugin *TestPlugin) GetDetachedVolumes() map[string][]string {
-	plugin.pluginLock.RLock()
-	defer plugin.pluginLock.RUnlock()
-	ret := make(map[string][]string)
-	for nodeName, volumeList := range plugin.detachedVolumeMap {
-		ret[nodeName] = make([]string, len(volumeList))
-		copy(ret[nodeName], volumeList)
-	}
-	return ret
-}
-
-func CreateTestPlugin() []volume.VolumePlugin {
-	attachedVolumes := make(map[string][]string)
-	detachedVolumes := make(map[string][]string)
-	return []volume.VolumePlugin{&TestPlugin{
-		ErrorEncountered:  false,
-		attachedVolumeMap: attachedVolumes,
-		detachedVolumeMap: detachedVolumes,
-		pluginLock:        &sync.RWMutex{},
-	}}
-}
-
-// Attacher
-type testPluginAttacher struct {
-	ErrorEncountered  *bool
-	attachedVolumeMap map[string][]string
-	pluginLock        *sync.RWMutex
-}
-
-func (attacher *testPluginAttacher) Attach(spec *volume.Spec, nodeName types.NodeName) (string, error) {
-	attacher.pluginLock.Lock()
-	defer attacher.pluginLock.Unlock()
-	if spec == nil {
-		*attacher.ErrorEncountered = true
-		return "", fmt.Errorf("Attach called with nil volume spec")
-	}
-	attacher.attachedVolumeMap[string(nodeName)] = append(attacher.attachedVolumeMap[string(nodeName)], spec.Name())
-	return spec.Name(), nil
-}
-
-func (attacher *testPluginAttacher) VolumesAreAttached(specs []*volume.Spec, nodeName types.NodeName) (map[*volume.Spec]bool, error) {
-	return nil, nil
-}
-
-func (attacher *testPluginAttacher) WaitForAttach(spec *volume.Spec, devicePath string, pod *v1.Pod, timeout time.Duration) (string, error) {
-	attacher.pluginLock.Lock()
-	defer attacher.pluginLock.Unlock()
-	if spec == nil {
-		*attacher.ErrorEncountered = true
-		return "", fmt.Errorf("WaitForAttach called with nil volume spec")
-	}
-	fakePath := fmt.Sprintf("%s/%s", devicePath, spec.Name())
-	return fakePath, nil
-}
-
-func (attacher *testPluginAttacher) GetDeviceMountPath(spec *volume.Spec) (string, error) {
-	attacher.pluginLock.Lock()
-	defer attacher.pluginLock.Unlock()
-	if spec == nil {
-		*attacher.ErrorEncountered = true
-		return "", fmt.Errorf("GetDeviceMountPath called with nil volume spec")
-	}
-	return "", nil
-}
-
-func (attacher *testPluginAttacher) MountDevice(spec *volume.Spec, devicePath string, deviceMountPath string, _ volume.DeviceMounterArgs) error {
-	attacher.pluginLock.Lock()
-	defer attacher.pluginLock.Unlock()
-	if spec == nil {
-		*attacher.ErrorEncountered = true
-		return fmt.Errorf("MountDevice called with nil volume spec")
-	}
-	return nil
-}
-
-// Detacher
-type testPluginDetacher struct {
-	detachedVolumeMap map[string][]string
-	pluginLock        *sync.RWMutex
-}
-
-func (detacher *testPluginDetacher) Detach(volumeName string, nodeName types.NodeName) error {
-	detacher.pluginLock.Lock()
-	defer detacher.pluginLock.Unlock()
-	detacher.detachedVolumeMap[string(nodeName)] = append(detacher.detachedVolumeMap[string(nodeName)], volumeName)
-	return nil
-}
-
-func (detacher *testPluginDetacher) UnmountDevice(deviceMountPath string) error {
-	return nil
 }
